+++ src/Blueprint/Blueprint/Cmdlets/BlueprintAssignment/RemoveAzureRMBlueprintAssignment.cs
@@ -23,7 +23,7 @@ using ParameterSetNames = Microsoft.Azure.Commands.Blueprint.Common.BlueprintCon
 
 namespace Microsoft.Azure.Commands.Blueprint.Cmdlets
 {
-    [GenericBreakingChangeWithVersion("Blueprints and associated cmdlets are being deprecated with no direct replacement.", "", "")]
+    [GenericBreakingChangeWithVersion("Blueprints and associated cmdlets are being deprecated with an encouraged migration to Deployments Stacks.", "", "")]
     [Cmdlet(VerbsCommon.Remove, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "BlueprintAssignment", SupportsShouldProcess = true, DefaultParameterSetName = ParameterSetNames.BySubscriptionAndName), OutputType(typeof(PSBlueprintAssignment))]
     public class RemoveAzureRmBlueprintAssignment : BlueprintAssignmentCmdletBase
     {