+++ src/Blueprint/Blueprint/Cmdlets/BlueprintDefinition/SetAzureRMBlueprint.cs
@@ -22,7 +22,7 @@ using static Microsoft.Azure.Commands.Blueprint.Common.BlueprintConstants;
 
 namespace Microsoft.Azure.Commands.Blueprint.Cmdlets
 {
-    [GenericBreakingChangeWithVersion("Blueprints and associated cmdlets are being deprecated with no direct replacement.", "", "")]
+    [GenericBreakingChangeWithVersion("Blueprints and associated cmdlets are being deprecated with an encouraged migration to Deployments Stacks.", "", "")]
     [Cmdlet(VerbsCommon.Set, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "Blueprint", SupportsShouldProcess = true, DefaultParameterSetName = ParameterSetNames.UpdateBlueprintBySubscription), OutputType(typeof(PSBlueprint))]
     public class SetAzureRMBlueprint : BlueprintDefinitionCmdletBase
     {