+++ src/Blueprint/Blueprint/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Deprecation of Blueprint cmdlets.
 
 ## Version 0.4.4
 * Removed Microsoft.Azure.Management.Blueprint 0.20.7-preview dependencies